oauth: file separation, classes restructured, fix refresh error, model classes for clarity
